Layout Header inkl. Sprachauswahl; Services in BackendAdapter.js ausgelagert
