+++ ui/ClientApp/src/components/Constants.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+export const arrowClosed = (
+	<span className="arrow-closed" />
+)
+export const arrowOpen = (
+	<span className="arrow-open" />
+)
+
+
+
