+++ ui/ClientApp/src/index.scss
@@ -24,15 +24,23 @@ body {
 	padding-top: 0;
 	margin: 0;
 	color: $dark-grey;
-  //background-color: bisque; 
-  background-color: $dark-white;
+  background-color: bisque; 
+  //background-color: $dark-white;
 	height: 100%;
   width: 100%;
   letter-spacing: 0.05em; 
 }
 
 header {
-	height: 20px;
+  display: block;
+  border-bottom: .0625rem solid $dark-grey;
+  // background: #fff;
+}
+
+.logos  {
+  max-width: 120rem;
+  margin: 0 auto;
+  padding-bottom: 0.5em;
 }
 
 footer {
